commit msg: python3

+++ to-jpg/to-jpg.py
@@ -3,11 +3,14 @@ from ftplib import FTP
 from PIL import Image
 from psd_tools import PSDImage
 
+mode = 'ADD'  # ADD | TEST | BULK
+skipftp = False
+QUALITY = 50
+IM_SIZE = 800
+
 savos = "/Users/thomasnieborowski/Desktop/SAVOS/"
 remote_img = 'public_html/sebartsvirtual/wp-content/uploads/img'
 
-mode = 'ADD' # ADD | TEST | BULK
-print('Processing Images in Mode: ' + mode)
 if mode == 'TEST':
     in_dir = "./in_img"
     out_dir = "./out_img"
@@ -17,65 +20,70 @@ elif mode == 'ADD':
 else:
     in_dir = savos + "IN_IMG"
     out_dir = savos + "OUT_IMG"
-
 save_dir = savos + "ORG_ADD_IN_IMG"
+print('Processing Images in Mode: ' + mode)
+print('Skipping FTP: ' + str(skipftp))
+print('Image Root Dir: ' + savos)
+print('input dir: ' + in_dir)
+print('output dir: ' + out_dir)
+print('Input files saved in : ' + save_dir)
+print('quality : ' + str(QUALITY))
+print('Max Size : ' + str(IM_SIZE))
 
-QUALITY = 50
-IM_SIZE = 800
+if not skipftp:
+    # save originally artist supplied files from previous run and clear in_dir.
+    for save in os.listdir(in_dir):
+        local_filename = os.path.join(in_dir, save)
+        os.rename(local_filename, os.path.join(save_dir, save) )
 
-# save originally artist supplied files from previous run and clear in_dir.
-for save in os.listdir(in_dir):
-    local_filename = os.path.join(in_dir, save)
-    os.rename(local_filename, os.path.join(save_dir, save) )
+    # FTP files to local
+    with open(os.path.join(savos, "savospw"), "r") as f1:
+        pw = f1.read().replace('\n', '').split(',')
 
-# FTP files to local
-with open(os.path.join(savos, "savospw"), "r") as f1:
-    pw = f1.read().replace('\n', '').split(',')
+    ftp = FTP(pw[0], pw[1], pw[2])
+    ftp.cwd(remote_img)
+    files = ftp.nlst()
 
-ftp = FTP(pw[0], pw[1], pw[2])
-ftp.cwd(remote_img)
-files = ftp.nlst()
+    for f2 in files:
+        if f2 == '.' or f2 == '..' or f2 == '.DS_Store':
+            continue
+        localf = os.path.join(in_dir, f2)
+        with open(localf, 'wb') as f3:
+            ftp.retrbinary('RETR '+f2, f3.write)
 
-for f2 in files:
-    if f2 == '.' or f2 == '..':
-        continue
-    localf = os.path.join(in_dir, f2)
-    with open(localf, 'wb') as f3:
-        ftp.retrbinary('RETR '+f2, f3.write)
+    for f4 in files:
+        if f4 == '.' or f4 == '..':
+            continue
+        ftp.delete(f4)
 
-for f4 in files:
-    if f4 == '.' or f4 == '..':
-        continue
-    ftp.delete(f4)
+    ftp.quit()
 
-ftp.quit()
-
+    for oldfile in os.scandir(out_dir):
+        os.remove(oldfile.path)
 
 # Resize and compress files
-for oldfile in os.scandir(out_dir):
-    os.remove(oldfile.path)
-
 for f4 in os.listdir(in_dir):
     org = os.path.join(in_dir, f4)
-    new_path = os.path.join(out_dir, f4) # TODO append out_dir last
+    newfile = f4
+    print('CONVERTing '+f4)
     if f4.endswith('.jpeg'):
-        newpath = new_path.replace('.jpeg', '.jpg')
+        newfile = newfile.replace('.jpeg', '.jpg')
         im = Image.open(org)
     elif f4.endswith('.jpg'):
         im = Image.open(org)
-        newpath = new_path
+        newfile = newfile
     elif f4.endswith('.png'):
-        newpath = new_path.replace('.png', '.jpg')
+        newfile = newfile.replace('.png', '.jpg')
         im = Image.open(org)
         im = im.convert('RGB')
     elif f4.endswith('.tif'):
-        newpath = new_path.replace('.tif', '.jpg')
+        newfile = newfile.replace('.tif', '.jpg')
         im = Image.open(org)
         im = im.convert('RGB')
     elif f4.endswith('.psd'):
         png = org.replace('.psd', '.png')
         os.system('psd-tools convert '+org+' '+png)
-        newpath = new_path.replace('.psd', '.jpg')
+        newfile = newfile.replace('.psd', '.jpg')
         im = Image.open(png)
         im = im.convert('RGB')
         os.remove(png)
@@ -83,8 +91,9 @@ for f4 in os.listdir(in_dir):
         print('ERROR: UNEXPECTED FILE EXTENSION: ' + f4)
         continue
     im.thumbnail((IM_SIZE, IM_SIZE))
+    newpath = os.path.join(out_dir, newfile)
     im.save(newpath, quality=QUALITY)
-    print('CONVERTed '+f4+' to '+newpath)
+    print('CONVERTed '+f4+' to '+newfile)
 
 print('Now upload precessed Images to Media Library:')
 print('Dashboard > Media > Add New > Select > ' + out_dir)
